+++ config/initializers/assets.rb
@@ -19,6 +19,12 @@ Rails.application.config.assets.precompile += %w(
   govuk_publishing_components/govuk-schema-placeholder-4x3.png
   govuk_publishing_components/govuk-schema-placeholder-16x9.png
   govuk_publishing_components/search-button.png
+  govuk_publishing_components/chevron-banner.svg
+  govuk_publishing_components/chevron-banner-focus.svg
+  govuk_publishing_components/chevron-banner-hover-border.svg
+  govuk_publishing_components/chevron-banner-hover.svg
+  govuk_publishing_components/chevron-banner-small.svg
+  govuk_publishing_components/chevron-banner-small-focus.svg
 )
 
 # GOV.UK Frontend assets