+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Ensure input styles are not overridden by static ([PR #1094](https://github.com/alphagov/govuk_publishing_components/pull/1094))
 * Fixes focus state spacing on 'extensive' related navigation links ([PR #1085](https://github.com/alphagov/govuk_publishing_components/pull/1085))
+* Add chevron banner component ([PR #1084](https://github.com/alphagov/govuk_publishing_components/pull/1084))
 
 ## 20.1.0